use double instead of float

+++ src/OpenCOVER/plugins/hlrs/Energy/lib/core/interfaces/IMovable.h
@@ -3,9 +3,9 @@
 
 namespace core::interface {
 struct Pos {
-    float x, y, z;
+    double x, y, z;
     Pos() : x(0.0f), y(0.0f), z(0.0f) {}
-    Pos(float x, float y, float z) : x(x), y(y), z(z) {}
+    Pos(double x, double y, double z) : x(x), y(y), z(z) {}
     Pos(const float xyz[3])
         : x(xyz[0])
         , y(xyz[1])
